ipam: Move helper functions out of clusterpool.go
This commit moves various functions used by the multipool.go
implementation out of clusterpool.go. This will allow us to remove the
clusterpool.go file in a subsequent commit.

This commit contains no functional changes.
