+++ pkg/ipam/pool.go
@@ -7,8 +7,11 @@ import (
 	"errors"
 	"fmt"
 	"net"
+	"strings"
 
 	"github.com/sirupsen/logrus"
+	"github.com/vishvananda/netlink"
+	"golang.org/x/sys/unix"
 
 	"github.com/cilium/cilium/pkg/defaults"
 	"github.com/cilium/cilium/pkg/ipam/service/ipallocator"
@@ -406,4 +409,61 @@ func (p *podCIDRPool) updatePool(podCIDRs []string) {
 	}
 
 	p.ipAllocators = newIPAllocators
+}
+
+func podCIDRFamily(podCIDR string) Family {
+	if strings.Contains(podCIDR, ":") {
+		return IPv6
+	}
+	return IPv4
+}
+
+// containsCIDR checks if the outer IPNet contains the inner IPNet
+func containsCIDR(outer, inner *net.IPNet) bool {
+	outerMask, _ := outer.Mask.Size()
+	innerMask, _ := inner.Mask.Size()
+	return outerMask <= innerMask && outer.Contains(inner.IP)
+}
+
+// cleanupUnreachableRoutes remove all unreachable routes for the given pod CIDR.
+// This is only needed if EnableUnreachableRoutes has been set.
+func cleanupUnreachableRoutes(podCIDR string) error {
+	_, removedCIDR, err := net.ParseCIDR(podCIDR)
+	if err != nil {
+		return err
+	}
+
+	var family int
+	switch podCIDRFamily(podCIDR) {
+	case IPv4:
+		family = netlink.FAMILY_V4
+	case IPv6:
+		family = netlink.FAMILY_V6
+	default:
+		return errors.New("unknown pod cidr family")
+	}
+
+	routes, err := netlink.RouteListFiltered(family, &netlink.Route{
+		Table: unix.RT_TABLE_MAIN,
+		Type:  unix.RTN_UNREACHABLE,
+	}, netlink.RT_FILTER_TABLE|netlink.RT_FILTER_TYPE)
+	if err != nil {
+		return fmt.Errorf("failed to fetch unreachable routes: %w", err)
+	}
+
+	var errs error
+	for _, route := range routes {
+		if !containsCIDR(removedCIDR, route.Dst) {
+			continue
+		}
+
+		err = netlink.RouteDel(&route)
+		if err != nil && !errors.Is(err, unix.ESRCH) {
+			// We ignore ESRCH, as it means the entry was already deleted
+			errs = errors.Join(errs, fmt.Errorf("failed to delete unreachable route for %s: %w",
+				route.Dst.String(), err),
+			)
+		}
+	}
+	return errs
 }
